Change PyCharm settings for Python 3
Previously I had used PyCharm in Python 2.x environments, and the string literal
colour was a yellowish.  As it turned out the colour for unicode string literals
(which is what all string literals are by default in Python 3) was a dark bluish
green which was hard to see.  Changed settings so that bytes and unicode string
literals are the same colour